feat: merge auth tests into AuthControllerTest
